import_key_material/requirements.txt
boto3==1.10.50
moto==1.3.14
cryptography==2.8
pyOpenSSL==19.1.0
mock==3.0.5
pytest==5.3.2
rsa==4.0
pycryptodome==3.9.4